third_party/WebKit/LayoutTests/fast/forms/form-submit-in-submit-event.html

<!DOCTYPE html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<iframe name="target"></iframe>
<form action="about:blank" onsubmit="this.submit(); return false;" target="target">
 <input name="t" value="text">
 <input type="submit" name="b" value="Next">
</form>

<script>
var test = async_test('submit() in submit event handler should not contains a value of a submit button.');
test.step(() => {
  var iframe = document.querySelector('iframe');
  iframe.addEventListener('load', test.step_func(() => {
    assert_true(iframe.contentWindow.location.href.endsWith('?t=text'));
    test.done();
  }), false);
  document.querySelector('input[type=submit]').click();
});
</script>
